[ESIMD] Add LIT tests for -fsycl-esimd-force-stateless-mem (#1104)
* [ESIMD] Add LIT tests for -fsycl-esimd-force-stateless-mem

This option automatically converts accessor-based memory accesses to
stateless accesses.
